// src/interfaces/user.ts
export interface IGithubUser {
  login: string;
  avatar_url: string;
  name: "Manuel Jaime";
  followers: number;
  following: number;
  html_url?: string;
  bio: string;
  company: string;
    location: "Luanda - Angola";
  //   blog: string;
  //   twitter_username?: string;
    starred_url: string;
  //   subscriptions_url: string;
  //   repos_url: string;
  //   email?: string;
  //   hireable: null;
    public_repos: number;
  //   public_gists: number;
}
// export interface IGithubUser {
//   avatar_url: "https://avatars.githubusercontent.com/u/187278126?v=4";
//   html_url: "https://github.com/manueljaime007";
//   followers_url: "https://api.github.com/users/manueljaime007/followers";
//   following_url: "https://api.github.com/users/manueljaime007/following{/other_user}";
//   starred_url: "https://api.github.com/users/manueljaime007/starred{/owner}{/repo}";
//   subscriptions_url: "https://api.github.com/users/manueljaime007/subscriptions";
//   repos_url: "https://api.github.com/users/manueljaime007/repos";
//   type: "User";
//   name: "Manuel Jaime";
//   company: "ITEL - Instituto de Telecomunicações";
//   blog: "https://manueljaime.vercel.app";
//   location: "Luanda - Angola";
//   email: null;
//   hireable: null;
//   bio: "IT student | Programmer";
//   twitter_username: null;
//   public_repos: 18;
//   public_gists: 0;
//   followers: 9;
//   following: 13;
// }